Merge develop into master (#49)
* fix(hh/forge): typechain support (#45)

Typechain uses a private method on the artifact
with an unsafe `any` typecast to get artifact paths.
To work with typechain, this method must exist until
typechain is updated.

The issue is here: https://github.com/dethcrypto/TypeChain/issues/544

* feat: `Artifacts.getBuildInfo` (#48)

Implement a naive `getBuildInfo` function. This is useful
for tooling that integrates more deeply with the hardhat ecosystem.
Note that this implementation is not 100% with the output
that the hardhat compiler toolchain outputs, but the output
is good enough to get more tooling to be compatible.