feat(hardhat-forge): add `foundry` config section
The `hardhat.config.ts` now can have a `foundry` section that
allows users to place config values in. This makes it so that
expensive operations like generating the build info can be placed
in the `hardhat.config.ts` and not in the `foundry.toml` so that
`forge build` can be ligher than `hardhat compile`.