fix: rebase develop on master (#61)
* Version Packages (#44)

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>

* fix(hh/forge): typechain support (#45) (#46)

Typechain uses a private method on the artifact
with an unsafe `any` typecast to get artifact paths.
To work with typechain, this method must exist until
typechain is updated.

The issue is here: https://github.com/dethcrypto/TypeChain/issues/544

Co-authored-by: Mark Tyneway <mark.tyneway@gmail.com>

* Version Packages (#47)

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>

* Merge develop into master (#49)

* fix(hh/forge): typechain support (#45)

Typechain uses a private method on the artifact
with an unsafe `any` typecast to get artifact paths.
To work with typechain, this method must exist until
typechain is updated.

The issue is here: https://github.com/dethcrypto/TypeChain/issues/544

* feat: `Artifacts.getBuildInfo` (#48)

Implement a naive `getBuildInfo` function. This is useful
for tooling that integrates more deeply with the hardhat ecosystem.
Note that this implementation is not 100% with the output
that the hardhat compiler toolchain outputs, but the output
is good enough to get more tooling to be compatible.

* Version Packages

* Merge develop into master (#55)

* fix(hh/forge): typechain support (#45)

Typechain uses a private method on the artifact
with an unsafe `any` typecast to get artifact paths.
To work with typechain, this method must exist until
typechain is updated.

The issue is here: https://github.com/dethcrypto/TypeChain/issues/544

* feat: `Artifacts.getBuildInfo` (#48)

Implement a naive `getBuildInfo` function. This is useful
for tooling that integrates more deeply with the hardhat ecosystem.
Note that this implementation is not 100% with the output
that the hardhat compiler toolchain outputs, but the output
is good enough to get more tooling to be compatible.

* hardhat-forge: use the correct source name (#51)

The relative path to the file from the root
of the project should be used as the source name
in the artifact.

Closes https://github.com/foundry-rs/hardhat/issues/36

This will fix tooling that depends on the source name:
https://github.com/defi-wonderland/smock/blob/27fe7a5f3c351714257cdc98204e282932c11e12/src/utils/storage.ts#L54

* test: explicitly set port for anvil (#53)

* test: explicitly set port for anvil

* chore: add changeset

Co-authored-by: Matthias Seitz <matthias.seitz@outlook.de>

* Version Packages (#56)

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>

* feat: `Artifacts.getBuildInfo` (#48)

Implement a naive `getBuildInfo` function. This is useful
for tooling that integrates more deeply with the hardhat ecosystem.
Note that this implementation is not 100% with the output
that the hardhat compiler toolchain outputs, but the output
is good enough to get more tooling to be compatible.

* hardhat-forge: use the correct source name (#51)

The relative path to the file from the root
of the project should be used as the source name
in the artifact.

Closes https://github.com/foundry-rs/hardhat/issues/36

This will fix tooling that depends on the source name:
https://github.com/defi-wonderland/smock/blob/27fe7a5f3c351714257cdc98204e282932c11e12/src/utils/storage.ts#L54

* test: explicitly set port for anvil (#53)

* test: explicitly set port for anvil

* chore: add changeset

* feat: better `BuildInfo` support (#59)

* feat(hardhat-forge): better `BuildInfo` support

Requires foundry with https://github.com/foundry-rs/foundry/commit/6ce16a477e69e58bbcc618341181d4096e3720a5.
Generates the debug files for easy access to the build info files
during hardhat scripts.

* feat(hardhat-forge): buildinfo tests

* feat(hardhat-forge): better `BuildInfo` support

Co-authored-by: Georgios Konstantopoulos <me@gakonst.com>
Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>
Co-authored-by: Matthias Seitz <matthias.seitz@outlook.de>