feat: configurable runSuper
Allow the user to configure `runSuper` in the
build task. Allowing the user to run this will
allow for more composability with other hardhat
based tooling but will slow down the build.

It was previously defaulted to true, but I found
in practice that its generally not necessary to
be true for just compilation. Things like `typechain`
will not automatically run afterwards, but `typechain`
doesn't currently work exactly the same as the foundry
`Artifacts` doesn't have the concept of `validArtifacts`
populated after building. We run `typechain` on its own
after compilation to get around this difference, so
always calling `runSuper` just slows down the build.