feat(hardhat-forge): allow skipping of tests
Add a new config option that allows users to skip
compilation of their tests. A test is defined
as a file that ends with a `.t.sol` extension.
This speeds up compilation quite a bit in the
optimism codebase, as compiling the tests
is the slowest part of the compilation pipeline.

If a user specifies the following in the `hardhat.config.ts`:

```js
  foundry: {
    skipTests: true
  }
```

Then the tests will not be compiled