feat(hardhat-forge): build info path config (#84)
Allow the `build_info_path` config option to be parsed
and passed through the hardhat plugin. It will correctly
create the debug files that point to the build info files
dynamically based on the configured build info path.

The tests that fetch the build info objects still pass, they
look up the path from the debug file to the build info file.