hardhat-forge: parse string metadata (#85)
* hardhat-forge: failing test

It should be generating metadata as an object,
not as a string. Its a JSON string that is double
stringified.

* hardhat-forge: parse build info metadata

Deserialize the build info compiler output
as an object because ethers-solc serializes
it as a string. Hardhat plugins expect it to
be an object and not a string.

* chore: add changeset

foundry-rs/hardhat-forge