hardhat-forge: handle case insensitive filesystem
The artifact paths and filenames are based on the names
of the files containing the smart contracts and the names of
the contracts inside of the files.

The npm package `true-case-path` has a bug that causes it to
ignore directories that are spelled the same but have alternate
casing.

You can see here that it eagerly returns if the case insensitive
regex hits:
https://github.com/Profiscience/true-case-path/blob/8a016e6a8be64c873aba414fbcdb4748e24dc796/index.js#L31

The problem is that it could hit multiple files and we need to check
each, not just the first one that it hits.

The problem is with `DSTest` and `forge-std`, where `DSTest` has
a `test.sol` and `forge-std` has a `Test.sol`. The regex hits
incorrectly and downstream consumers get passed the wrong data.

The fix simply looks to see if `true-case-path` has an error
(returns `null`) and then uses `fsExtra` to look directly at the
filesystem. If directly looking at the filesystem works, then
we know that the file is there and don't need to error.