hardhat-forge: ignore metadata files
When getting contract artifacts, json files are pulled out of
the artifacts directory. Since metadata files are now emitted by
default as of
https://github.com/foundry-rs/foundry/commit/ad91a4a962ddf2f3c3e340c018482851a5080f01,
extra files were being passed through the hh artifact generation
pipeline which broke things because the hardhat artifacts need
the fully qualified name which is currently pulled from the
ast in the compiler out. It could be possible to brute force
the fully qualified name but that would fail when 2 contracts
have a matching name. This filters out the metadata files
when getting the contract artifacts.